+++ test/com/hennrix/physics/geometry/TestPoint.java
@@ -1,6 +1,6 @@
 package com.hennrix.physics.geometry;
 
-import com.hennrix.math.interfaces.Point;
+import com.hennrix.math.model.Point;
 import com.hennrix.physics.Coordinates;
 
 public class TestPoint extends Point {
